+++ include/libfork/core.hpp
@@ -37,6 +37,7 @@
 #include "libfork/core/impl/promise.hpp"
 #include "libfork/core/impl/return.hpp"
 #include "libfork/core/impl/stack.hpp"
+#include "libfork/core/impl/unique_frame.hpp"
 #include "libfork/core/impl/utility.hpp"
 
 /**